Add read_message_with_nonce for unreliable transports
As mentioned in section 11.4 of the noise protocol whitepaper, using noise with
udp is possible, if the n values are tracked manually. This is already
possible by calling set_receiving_nonce explicitly before calling read_message.

Since the n values need to be tracked manually the internal cipherstate doesn't
need to be mutable anymore, which is the only reason the encrypt/decrypt
functions need a mutable borrow. Changing this to an immutable borrow allows
decryption with multiple threads.